Refactor evaluation to separate crate
