+++ backend/evaluator/Cargo.toml
@@ -1,7 +1,13 @@
 [package]
-name = "eka_ci_evaluator"
+name = "evaluator"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.5.20", features = ["derive"] }
+shared = { workspace = true }
+anyhow = { workspace = true }
+async-trait = "0.1"
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { version = "1.41.0", features = ["full", "process", "io-util"] }
+tracing = { workspace = true }
