--- backend/evaluator/src/cli/mod.rs
@@ -1,6 +1,0 @@
-use clap::Parser;
-
-/// Simple program to greet a person
-#[derive(Parser, Debug)]
-#[command(version, about, long_about = None)]
-pub(crate) struct Args {}
