+++ backend/evaluator/src/lib.rs
@@ -1,0 +1,12 @@
+pub mod nix_utils;
+pub mod service;
+pub mod traits;
+pub mod types;
+pub mod utils;
+
+// Re-export commonly used items at the crate root for convenience
+pub use traits::{EvalDatabase, EvalMetricsCollector, NullMetrics};
+pub use types::{
+    DrvInfo, DrvOutput, DrvPackageMetadata, NixEvalDrv, NixEvalError, NixEvalItem, NixEvalMeta,
+};
+pub use utils::{pname_from_name, version_from_name};
