--- backend/evaluator/src/main.rs
@@ -1,6 +1,0 @@
-mod cli;
-use clap::Parser;
-
-fn main() {
-    cli::Args::parse();
-}
