+++ backend/evaluator/src/nix_utils/mod.rs
@@ -1,0 +1,6 @@
+pub mod commands;
+pub mod size;
+
+// Re-export commonly used functions
+pub use commands::{drv_references, drv_requisites, get_drv_outputs, is_drv_cached, output_references};
+pub use size::{format_size, get_closure_sizes, get_output_sizes};
