+++ backend/evaluator/src/service/mod.rs
@@ -1,0 +1,7 @@
+pub mod jobs;
+
+// Re-export key functions and types
+pub use jobs::{
+    process_nix_eval_output, run_nix_eval_jobs, ConsumeOutcome, Truncation,
+    NIX_EVAL_JOBS_MAX_ENTRIES, NIX_EVAL_JOBS_MAX_LINE_BYTES, NIX_EVAL_JOBS_MAX_STDOUT_BYTES,
+};
