+++ backend/evaluator/src/types/mod.rs
@@ -1,0 +1,9 @@
+pub mod derivation_show;
+pub mod nix_eval_jobs;
+
+// Re-export commonly used types
+pub use derivation_show::{DrvInfo, DrvOutput, drv_output};
+pub use nix_eval_jobs::{
+    DrvPackageMetadata, HomepageField, LicenseEntry, LicenseField, MaintainerEntry, NixEvalDrv,
+    NixEvalError, NixEvalItem, NixEvalMeta,
+};
