+++ backend/server/src/gitea/actions.rs
@@ -164,7 +164,7 @@ pub async fn fail_ci_eval_job(
     client: &GiteaClient,
     ci_info: &GiteaCIInfo,
     job_name: &str,
-    errors: &[crate::nix::nix_eval_jobs::NixEvalError],
+    errors: &[crate::nix::NixEvalError],
 ) -> Result<i64> {
     debug!(
         "Creating failed CI eval job check run for job '{}' on commit {} with {} errors",
