+++ backend/server/src/nix/mod.rs
@@ -1,7 +1,20 @@
-pub mod derivation_show;
-pub mod jobs;
-pub mod nix_eval_jobs;
-pub mod size;
+// Re-export evaluator types and functions for backward compatibility
+pub use evaluator::types::{
+    derivation_show, DrvInfo, DrvOutput, DrvPackageMetadata, NixEvalDrv, NixEvalError, NixEvalItem,
+    NixEvalMeta,
+};
+pub use evaluator::nix_utils::{
+    self as size, drv_references, drv_requisites, format_size, get_closure_sizes,
+    get_drv_outputs, get_output_sizes, is_drv_cached, output_references,
+};
+pub use evaluator::service::jobs::{
+    process_nix_eval_output, Truncation, NIX_EVAL_JOBS_MAX_ENTRIES,
+    NIX_EVAL_JOBS_MAX_LINE_BYTES, NIX_EVAL_JOBS_MAX_STDOUT_BYTES,
+};
+pub use evaluator::utils::{pname_from_name, version_from_name};
+
+// Server-specific wrapper for eval jobs
+mod jobs;
 
 use std::collections::{HashMap, HashSet};
 use std::num::NonZeroUsize;
@@ -213,7 +226,7 @@ impl EvalService {
         use tokio::task::JoinSet;
 
         debug!("Traversing drv tree for {}", drv_path);
-        let drvs: Vec<DrvId> = drv_requisites(drv_path).await?;
+        let drvs: Vec<DrvId> = drv_requisites_as_ids(drv_path).await?;
         let mut drv_map = self.drv_map.lock().await;
         let new_drvids: Vec<DrvId> = drvs
             .into_iter()
@@ -448,141 +461,20 @@ pub async fn dry_run_realise(drv: &DrvId) -> Result<DryRunReport> {
     Ok(DryRunReport::parse(&stderr))
 }
 
-/// Retreive the requisites of a drv. This is a global list of all direct
-/// and transitive drvs
-pub(crate) async fn drv_requisites(drv_path: &str) -> Result<Vec<DrvId>> {
+// The following functions are now imported from evaluator and re-exported above:
+// - drv_requisites: Returns Vec<String> of drv paths
+// - drv_references: Returns Vec<String> of direct drv dependencies
+// - output_references: Returns Vec<String> of runtime references for an output
+// - get_drv_outputs: Returns HashMap<String, String> of output names to paths
+
+// Helper wrapper to convert drv_requisites result to Vec<DrvId> for server use
+pub(crate) async fn drv_requisites_as_ids(drv_path: &str) -> Result<Vec<DrvId>> {
     use std::str::FromStr;
-
-    let output = Command::new("nix-store")
-        .args(["--query", "--requisites", drv_path])
-        .output()
-        .await?
-        .stdout;
-    let drv_str = String::from_utf8(output)?;
-
-    let drvs = drv_str
-        .lines()
-        // drv requisites can include "inputSrcs" which are not inputDrvs
-        // but rather files which were added to the nix store through
-        // path literals or `nix-store --add`
-        .filter(|x| x.ends_with(".drv"))
-        .map(DrvId::from_str)
+    let drvs = drv_requisites(drv_path).await?;
+    drvs.into_iter()
+        .map(|s| DrvId::from_str(&s))
         .collect::<Result<Vec<DrvId>, _>>()
-        .context("failed to parse drv path from nix-store --requisites output")?;
-
-    Ok(drvs)
-}
-
-// Retreive the direct dependencies of a drv
-pub async fn drv_references(drv_path: &str) -> Result<Vec<String>> {
-    let output = Command::new("nix-store")
-        .args(["--query", "--references", drv_path])
-        .output()
-        .await?
-        .stdout;
-    let drv_str = String::from_utf8(output)?;
-
-    let drvs = drv_str
-        .lines()
-        // drv references can include "inputSrcs" which are not inputDrvs
-        // but rather files which were added to the nix store through
-        // path literals or `nix-store --add`
-        .filter(|x| x.ends_with(".drv"))
-        .map(|x| x.to_string())
-        .collect::<Vec<String>>();
-
-    Ok(drvs)
-}
-
-/// Retrieve the runtime references of an output path (retained dependencies)
-///
-/// This queries what store paths are actually referenced by a built output,
-/// which represents the true runtime dependencies (what ends up in the closure).
-///
-/// # Arguments
-/// * `output_path` - Full store path to a built output (e.g., `/nix/store/hash-name`)
-///
-/// # Returns
-/// Vector of full store paths that are runtime dependencies
-pub async fn output_references(output_path: &str) -> Result<Vec<String>> {
-    let output = Command::new("nix-store")
-        .args(["--query", "--references", output_path])
-        .output()
-        .await?
-        .stdout;
-    let refs_str = String::from_utf8(output)?;
-
-    let refs = refs_str
-        .lines()
-        .filter(|x| !x.is_empty())
-        .map(|x| x.trim().to_string())
-        .collect::<Vec<String>>();
-
-    Ok(refs)
-}
-
-/// Get the outputs of a derivation with their names
-///
-/// Uses `nix derivation show` to get structured information about a derivation's outputs.
-///
-/// # Arguments
-/// * `drv_path` - Path to the .drv file
-///
-/// # Returns
-/// HashMap mapping output names (e.g., "out", "dev", "doc") to their store paths
-pub async fn get_drv_outputs(drv_path: &str) -> Result<HashMap<String, String>> {
-    let output = Command::new("nix")
-        .args(["derivation", "show", drv_path])
-        .output()
-        .await
-        .context("Failed to execute nix derivation show")?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        anyhow::bail!("nix derivation show failed for {}: {}", drv_path, stderr);
-    }
-
-    let json_str = String::from_utf8(output.stdout)?;
-    let drv_output: derivation_show::DrvOutput =
-        serde_json::from_str(&json_str).context("Failed to parse nix derivation show output")?;
-
-    // The output is a map with the drv path as key, get the first (and only) value
-    let drv_info = drv_output
-        .drvs
-        .into_values()
-        .next()
-        .context("No derivation info found in output")?;
-
-    // Extract outputs map
-    let outputs_map: HashMap<String, String> = if let Some(outputs) = drv_info.outputs {
-        outputs
-            .into_iter()
-            .map(|(name, info)| (name, info.path))
-            .collect()
-    } else {
-        // Fallback: if no outputs field, assume single "out" output
-        // Try using nix-store --query --outputs as fallback
-        let fallback_output = Command::new("nix-store")
-            .args(["--query", "--outputs", drv_path])
-            .output()
-            .await?;
-
-        if fallback_output.status.success() {
-            let paths = String::from_utf8(fallback_output.stdout)?;
-            let first_path = paths.lines().next().unwrap_or("").trim();
-            if !first_path.is_empty() {
-                let mut map = HashMap::new();
-                map.insert("out".to_string(), first_path.to_string());
-                map
-            } else {
-                HashMap::new()
-            }
-        } else {
-            HashMap::new()
-        }
-    };
-
-    Ok(outputs_map)
+        .context("failed to parse drv path from requisites output")
 }
 
 // fn graph_line_to_drvids(drv_line: &str) -> Result<(DrvId, DrvId)> {
